// frontend/src/app/models/user/social-media-link.ts
/**
 * Ссылки на социальные сети
 * com.shtrokfm.quadrober.model.SocialMediaLink
 */
export interface SocialMediaLink {
  url: string;
  name: string;
//  TODO Возможно добавить тип социальной сети,
//   чтобы отображать красивее ссылки
//  private String type;
}